% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_bugzilla_key}
\alias{create_bugzilla_key}
\alias{check_authentication}
\title{Authentication}
\usage{
create_bugzilla_key(host)

check_authentication(key = Sys.getenv("R_BUGZILLA"), host)
}
\arguments{
\item{host}{URL of the bugzilla instance if missing the R bugzilla is assumed.}

\item{key}{API key name to check. By default R_BUGZILLA.}
}
\value{
TRUE invisibly if the actions are performed.
}
\description{
Obtain an API key or check if it can fine one that works for this host and
the version of the bugzilla provided works with this package.
}
\examples{
check_authentication()
}
\seealso{
set_key() if you want to temporary use an API key.
}
