% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugRzilla.R
\docType{package}
\name{bugRzilla}
\alias{bugRzilla}
\title{bugRzilla: A package to interact with Bugzilla.}
\description{
Interact as a user with a bugzilla site, by default R-project.
Download data associated with bugs, comments, attachments, history or post a
new bug.
}
\details{
When reporting a bug we hope to get something fixed for all and help the maintainers.
However, it also means more work for maintainers, so try to minimize their work.
Try to minimize the work maintainers need to do by following \href{https://bugs.r-project.org/bugzilla/page.cgi?id=bug-writing.html}{Bugzilla recommendations}.
}
\section{Reporting R bugs}{

If you want to report bugs in R itself consider reading the brief section
R-FAQ has about \href{https://mac.r-project.org/man/R-FAQ.html#R-Bugs}{bugs}.
Consider if this has been reported on the mailing list (R-devel) or on the
issue tracker itself.
}

