% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attachment.R
\name{get_attachment}
\alias{get_attachment}
\title{Retrieve information about attachments}
\usage{
get_attachment(issue, attachment, host)
}
\arguments{
\item{issue}{Number of the issue.}

\item{attachment}{A numeric ID of the attachment.}

\item{host}{URL of the bugzilla instance if missing the R bugzilla is assumed.}
}
\value{
A data.frame with the information available.
}
\description{
Given an issue or an attachment
}
\examples{
get_attachment(issue = 1)
get_attachment(attachment = 1)
}
