% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bug.R
\name{get_bug}
\alias{get_bug}
\title{Get information of issues}
\usage{
get_bug(issue, host)
}
\arguments{
\item{issue}{Number of the issue.}

\item{host}{URL of the bugzilla instance if missing the R bugzilla is assumed.}
}
\value{
A data.frame with all the information available and ordered by comment.
}
\description{
Retrieve data from the issues posted.
}
\examples{
gi <- get_bug(1)
}
