% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comment}
\alias{get_comment}
\title{Retrieve comment}
\usage{
get_comment(issue, comment, host)
}
\arguments{
\item{issue}{Number of the issue.}

\item{comment}{A character with the id of the comment eg. "1".}

\item{host}{URL of the bugzilla instance if missing the R bugzilla is assumed.}
}
\value{
All the available information.
}
\description{
Get comments from an issue or by id.
}
