% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{Get fields}
\usage{
get_fields(host)
}
\arguments{
\item{host}{URL of the bugzilla instance if missing the R bugzilla is assumed.}
}
\value{
A data.frame with three columns, Name of the fields, if it is
mandatory and the values it can take.
}
\description{
Retrieve possible fields to fill and their values if they are from a closed list.
}
\examples{
fields <- get_fields()
}
