% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bug_history.R
\name{get_history}
\alias{get_history}
\title{Provide information about the history of an issue.}
\usage{
get_history(issue, host, new_since = NULL)
}
\arguments{
\item{issue}{Number of the issue.}

\item{host}{URL of the bugzilla instance if missing the R bugzilla is assumed.}

\item{new_since}{A character with a Date in YYYY-MM-DD format.}
}
\description{
See what has changed of an issue.
}
\examples{
get_history(issue = 1)
}
