% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_bug.R
\name{post_bug}
\alias{post_bug}
\title{Open an issue}
\usage{
post_bug(description, summary, component, ..., version, product, host)
}
\arguments{
\item{...}{Named arguments passed to the API \href{https://bugzilla.readthedocs.io/en/latest/api/core/v1/bug.html}{check documentation}}
}
\value{
The ID of the issue.
}
\description{
Requires an user and an API key.
}
