% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{post_comment}
\alias{post_comment}
\title{Post a comment}
\usage{
post_comment(issue, comment, is_markdown = TRUE, host, key)
}
\arguments{
\item{issue}{A numeric value of the issue ID where the comment should be
posted.}

\item{comment}{The text you want to post.}

\item{is_markdown}{A logical value saying if it is markdown or not.}

\item{host}{URL of the bugzilla instance if missing the R bugzilla is assumed.}

\item{key}{API key name to check. By default R_BUGZILLA.}
}
\value{
The ID of the comment posted.
}
\description{
Add information to a bug report.
}
\seealso{
\code{post_bug()}, \code{get_comment()}
See the API reference: \url{https://bugzilla.readthedocs.io/en/latest/api/core/v1/comment.html#create-comments}
}
