% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_bugzilla_key}
\alias{create_bugzilla_key}
\alias{authentication}
\alias{check_authentication}
\title{Authentication}
\usage{
create_bugzilla_key(host = "https://bugs.r-project.org/bugzilla/")

check_authentication(
  key = Sys.getenv("R_BUGZILLA"),
  host = "https://bugs.r-project.org/bugzilla/"
)
}
\arguments{
\item{host}{URL of the bugzilla instance.}

\item{key}{API key to check.}
}
\value{
TRUE invisibly if the actions are performed.
}
\description{
Obtain an API key or check if it is working.
}
