% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_bugzilla_key}
\alias{create_bugzilla_key}
\title{Authentication}
\usage{
create_bugzilla_key(host)
}
\arguments{
\item{host}{URL of the bugzilla instance.}

\item{key}{API key to check.}
}
\value{
TRUE invisibly if the actions are performed.
}
\description{
Obtain an API key or check if it is working.
}
