% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue.R
\name{get_issue}
\alias{get_issue}
\title{Get information of a single issue}
\usage{
get_issue(issue, host)
}
\arguments{
\item{issue}{Number of the issue.}

\item{host}{URL of the bugzilla instance if missing the R bugzilla is assumed.}
}
\value{
A data.frame with all the information available and ordered by comment.
}
\description{
Retrieve data of an issue from the xml file associated with it (Doesn't use
the API).
}
\examples{
gi <- get_issue(1)
}
\seealso{
get_bug() for the equivalent using the API.
}
