% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue.R
\name{get_issue}
\alias{get_issue}
\title{Get information of a single issue}
\usage{
get_issue(issue, url = "https://bugs.r-project.org/bugzilla/")
}
\arguments{
\item{issue}{Number of the issue}

\item{url}{Url of the website.}
}
\description{
Get information of a single issue
}
