% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user}
\alias{get_user}
\title{User information}
\usage{
get_user(ids, names = NULL, host)
}
\arguments{
\item{ids}{Id of the users.}

\item{names}{Names of the users.}

\item{host}{URL of the bugzilla instance if missing the R bugzilla is assumed.}
}
\value{
Information about the users
}
\description{
Retrieve information about users. Requires authentication.
}
\examples{
gu <- get_user(1)
}
