% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{post_comment}
\alias{post_comment}
\title{Post a comment}
\usage{
post_comment(issue, comment, is_markdown, host)
}
\arguments{
\item{issue}{Number of the issue.}

\item{comment}{A character with the id of the comment eg. "1".}

\item{is_markdown}{A logical value saying if it is markdown or not.}

\item{host}{URL of the bugzilla instance if missing the R bugzilla is assumed.}
}
\value{
The ID of the comment posted.
}
\description{
Add information to a bug report.
}
