% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_key}
\alias{set_key}
\title{Set key to use}
\usage{
set_key(key = Sys.getenv("R_BUGZILLA"))
}
\arguments{
\item{key}{The API key you want to use.}
}
\value{
TRUE
}
\description{
If you don't want to have the API key on .Renviron or you want to use another
API key for bugzilla you can set it up this way. It will store the key in the
envirnoment which will only work for this R session.
}
